+++ API/src/models/Addresses.js
@@ -1,0 +1,11 @@
+import Query from "./Query.js";
+
+class Addresses {
+  static async getAllAddresses() {
+    const query = `SELECT * FROM addresses`;
+    const response = await Query.run(query);
+    return response;
+  }
+}
+
+export default Addresses;
